Update sample script to respond the new format.
